extract world removal class
