+++ Client/sdk/game/CGame.h
@@ -67,6 +67,7 @@ class CWeaponStatManager;
 class CWeather;
 class CWorld;
 class CIplStore;
+class CBuildingRemoval;
 enum eEntityType;
 enum ePedPieceTypes;
 
@@ -147,6 +148,7 @@ public:
     virtual CWeaponStatManager*       GetWeaponStatManager() = 0;
     virtual CPointLights*             GetPointLights() = 0;
     virtual CColStore*                GetCollisionStore() = 0;
+    virtual CBuildingRemoval*         GetBuildingRemoval() = 0;
 
     virtual CWeaponInfo* GetWeaponInfo(eWeaponType weapon, eWeaponSkill skill = WEAPONSKILL_STD) = 0;
     virtual CModelInfo*  GetModelInfo(DWORD dwModelID, bool bCanBeInvalid = false) = 0;
